refactor: move suggestion caching and channel-type logic into the app module
The suggestion domain (a SkyBlock-Nerds concept) was baked into discord-framework.
Move Suggestion and SuggestionCache into app, extract the forum/thread/channel
suggestion-type resolution out of the framework's DiscordUtils into a new
app SuggestionTypeResolver, and move the review-request-channel lookup out of the
framework's ChannelCache into SuggestionCommandUtils. Config reads in the moved
code now target NerdBotConfig. After this, discord-framework no longer references
suggestions or their config. Behaviour is unchanged.
